requirements.txt
openai-whisper
gTTS
sounddevice
numpy
scipy
pygame
google-generativeai
python-dotenv
groq
coqui-tts
torch
fastapi
uvicorn
jinja2
websockets
pydub
soundfile
aiofiles
python-multipart
livekit
livekit-agents
livekit-api
